fixed warning cc2530 conversation of uint16_t type

+++ os/net/mac/framer/framer-802154.c
@@ -114,7 +114,7 @@ create_frame(int do_create)
 
     LOG_INFO("Out: %2X ", params.fcf.frame_type);
     LOG_INFO_LLADDR((const linkaddr_t *)params.dest_addr);
-    LOG_INFO_(" %d %u (%u)\n", hdr_len, packetbuf_datalen(), packetbuf_totlen());
+    LOG_INFO_(" %d %u (%u)\n", hdr_len, packetbuf_datalen(), (unsigned int) packetbuf_totlen());
 
     return hdr_len;
   } else {
@@ -300,7 +300,7 @@ parse(void)
     LOG_INFO_LLADDR(packetbuf_addr(PACKETBUF_ADDR_SENDER));
     LOG_INFO_(" ");
     LOG_INFO_LLADDR(packetbuf_addr(PACKETBUF_ADDR_RECEIVER));
-    LOG_INFO_(" %d %u (%u)\n", hdr_len, packetbuf_datalen(), packetbuf_totlen());
+    LOG_INFO_(" %d %u (%u)\n", hdr_len, packetbuf_datalen(), (unsigned int) packetbuf_totlen());
 
     return hdr_len;
   }
